--- frontend/db.json
@@ -1,14 +1,0 @@
-{
-  "posts": [
-    {
-      "title": "Stalin ",
-      "author": "Cocadas caseiras ",
-      "id": 21
-    },
-    {
-      "title": "Gaby",
-      "author": "brincos",
-      "id": 23
-    }
-  ]
-}
