--- frontend/src/Apprutas.css
@@ -1,90 +1,0 @@
-body {
-  margin: 0;
-  /*display: flex;*/
-  align-items: center;
-  text-align: center;
-  /*background-color: #e0f6fc;*/
-  /* background: url("pasta/pexels-lukas-574071tres.jpg")*/
-}
-
-.centrasobrenos {
-  align-items: center;
-  display: flex;
-  text-align: center;
-  align-items: center;
-  justify-content: center;
-}
-
-.App-logo {
-  height: 12vmin;
-  padding: 0 1rem;
-  float: left;
-  animation: none;
-  max-width: 100%;
-  max-height: 100%;
-}
-
-/*@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
-}*/
-
-.presenta_css {
-  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
-    'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
-    sans-serif;
-  margin: 0 auto;
-  width: 40vw;
-  height: 40vh;
-  padding: 30px;
-  border: 6px solid black;
-  border-radius: 4em;
-  /*background-color: #f7f7f7;*/
-  margin-top: 4em;
-}
-
-.presenta_css h3 {
-  font-size: 1.6rem;
-  padding: 0;
-  margin-bottom: 10px;
-  margin-top: 0px;
-}
-
-.presenta_css form {
-  height: 90%;
-  font-size: 1.2rem;
-  padding: 0;
-  margin: 0 auto;
-  display: flex;
-  flex-direction: column;
-  font-family: system-ui, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto,
-    Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
-  text-align: start;
-}
-
-.controls {
-  padding: 5px;
-  border-radius: 4px;
-  height: 20px;
-  border: 0.01rem solid black;
-  font-size: 1rem;
-  margin-bottom: 0.5rem;
-}
-
-.button {
-  margin-top: 1rem;
-  background-color: #ca932b;
-  color: black;
-  font-weight: bolder;
-  font-size: 1.3rem;
-  border-radius: 0.4rem;
-  border: 0.1rem solid black;
-}
-
-#message {
-  height: 20rem;
-}
