+++ frontend/src/Screens/SobreNos.js
@@ -1,24 +1,22 @@
 import React from "react";
-
-import Carroulsel from "./Carroulsel";
+import Carousel from "./Carousel";
 import CardDinamic from "./CardDinamic"
-import "./App.css";
+import { Helmet } from "react-helmet-async";
 
 function SobreNos() {
   return (
     <div className="SobreNos">
-   
-
-      {/*<h1> CAMINHOS SUSTENTAVEIS</h1>*/}
+      <Helmet>
+        <title>Sobre Nós</title>
+      </Helmet>
       <CardDinamic />
       <h1>SobreNós</h1>
       <p>
-      Somos Migrantes Venezuelanos, uma equipe formada pela igualdade de gênero
-      </p> 
+        Somos Migrantes Venezuelanos, uma equipe formada pela igualdade de gênero
+      </p>
       <p>Composta por duas mulheres e dois homens com idades diferentes. Temos de 23 a 61 anos de idade. Essa diversidade nos permite expressar a necessidade de apoio à resiliência no Brasil a partir de diferentes perspectivas.
       </p>
-      <Carroulsel />
-   
+      <Carousel />
     </div>
   );
 }
