--- frontend/src/componentes/App.css
@@ -1,65 +1,0 @@
-.App {
-  text-align: center;
-}
-
-.Empleabilidade ul li {
-  list-style: none;
-  color: white;
-}
-.Empleabilidade ul li a{
-  list-style: none;
-  color: white;
-}
-
-.Empleabilidade {
-  background-color: rgb(10, 10, 221);
-  color:white
-}
-.Footer {
-
-  background-color: rgb(160, 124, 5);
-}
-.App-logo {
-  height: 40vmin;
-  pointer-events: none;
-}
-
-@media (prefers-reduced-motion: no-preference) {
-  .App-logo {
-    animation: App-logo-spin infinite 20s linear;
-  }
-}
-
-.App-header {
-  background-color: #282c34;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
-}
-
-.App-link {
-  color: #61dafb;
-}
-
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
-}
-
-.App-logo {
-  height: 12vmin;
-  padding: 0 1rem;
-  float: left;
-  animation: none;
-  max-width: 100%;
-  max-height: 100%;
-  position: relative;
-}
