--- frontend/src/componentes/Açouge.js
@@ -1,1 +1,0 @@
-
