--- frontend/src/componentes/Estadosba.js
@@ -1,77 +1,0 @@
-import React from "react";
-import "./App.css";
-import '../Apprutas.css';
-
-function Estadosba() {
-  return (
-    <section>
-      <div className="Container">
-        <h1>Empleabilidad</h1>
-        <p>
-          Caminos Sustentaves atendera dos areas fundamentales dentro de sus
-          funciones.
-        </p>
-        <li>Empleabilidad</li>
-        <li>Emprendurismo</li>
-
-        <img src="https://img.freepik.com/foto-gratis/cerrar-apreton-manos-gente-negocios_23-2148825865.jpg?w=740&t=st=1684693165~exp=1684693765~hmac=8579f7fe5c2623d9db3f7ad6d4716586f8b9b88be3aa8d2ad21162b96af162f5" alt=""/>
-        <img src="https://img.freepik.com/foto-gratis/sastre-mujer-trabajando-tela-cuero_1303-23022.jpg?w=740&t=st=1684693025~exp=1684693625~hmac=6455d250f8fda8013b01a487d709656a54c0507ef523c6526b75488993c89e54" alt=""/>
-      </div>
-
-      <div className="Empleabilidade">
-      <h1> Você está procurando por oportunidades de trabalho no Brasil?</h1>
-        <h2>“Nós damos alguns passos necessários para realizar seu sonho de um emprego</h2>
-        <p>
-        Nessa área de atenção, forneceremos conselhos sobre conhecimentos básicos e o 
-        </p><p>ABC de aspectos fundamentais que um migrante deve seguir para se adaptar à sua nova realidade migratória no Brasil. para se adaptar à sua nova realidade migratória no Brasil.
-        </p>
-        <ul>
-          <li>
-
-            <a href="https://conectesus-paciente.saude.gov.br/login">SAUDE</a>
-          </li>
-          <li>
-
-            <a href="https://help.unhcr.org/brazil/es/vivir-en-brasil/educacion/">
-              Educacion
-            </a>
-          </li>
-          <li>
-
-            <a href="https://www.gov.br/trabalho-e-previdencia/pt-br/servicos">
-              Carteira de Trabalho Digital
-            </a>
-          </li>
-          <li>
-            <a href="https://www.sebrae.com.br/sites/PortalSebrae/cursosonline"> Educacion
-            </a>
-          </li>
-        </ul>
-        <img src="https://img.freepik.com/foto-gratis/sonriente-chica-asiatica-barista-haciendo-cafe-sosteniendo-taza-delantal-uniforme-cafe-pie-cerca-cuenta_1258-130464.jpg?size=626&ext=jpg&uid=R68575601&ga=GA1.2.1328558379.1681791155&semt=sph" width="740 px" height= "493px" alt=""/>
-      <img src="https://img.freepik.com/foto-gratis/trabajador-supermercado-suministrando-alimentos-al-departamento-frutas_342744-1075.jpg?w=740&t=st=1684710070~exp=1684710670~hmac=b640388932d773d9cd8686a0bda3f5e17c76257707d4e4aa5d219573ece27102" alt=""/>
-      </div>
-
-      <div className="Emprendedurismo">
-<h1>Empreendedorismo: desde a ideia, manual de negócios, contabilidade básica e marketing, é por isso que temos nossa seção de empreendedorismo.</h1>
-        <img src="https://img.freepik.com/vector-gratis/concepto-innovacion-dibujado-mano-bombilla_23-2149152350.jpg?w=740&t=st=1684697808~exp=1684698408~hmac=768a436c042279a7ddb81a30fca341430409d98ae5cb1d533b80f181da2393f1"  width="300px" alt=""/>
-        < img src="https://img.freepik.com/vector-gratis/coleccion-artistas-trabajo_23-2148228987.jpg?w=740&t=st=1684698499~exp=1684699099~hmac=765f9f1536ddf9886de7bde22b588929706e0a41fb3ce7c6bc79d2fe40742aff" width="300px" alt=""/>
-        <br/>
-        < img src=" https://img.freepik.com/vector-gratis/concepto-ilustrado-contra-reembolso_23-2148789201.jpg?w=740&t=st=1684698729~exp=1684699329~hmac=f887b9bd4616e2c0991d00aa69fd10d936fcd5f2f1e81ced6048703a5e446afb" width= "300px" alt=""/>
-        <img src="https://img.freepik.com/vector-premium/compre-cartel-que-estamos-abiertos_23-2148550415.jpg?w=740"width="500px " alt=""/>
-
-      </div>
-      <div className="Estados">
-
-  <h1>POPULAÇÃO DE INTERESSE POR TERRITÓRIO NO BRASIL</h1>
-  <img src="https://journalabroad.com/wp-content/uploads/2018/04/recife-beach-sand-view.jpg" width= "1200px" height="400px" alt=""></img>
-<h1>Nordeste</h1>
-<article>O Nordeste, onde destacaremos nosso trabalho na Bahia, em Pernanbuco e no Rio Grande do Norte, concentra mais da metade dos extremamente pobres do Brasil, 53,2%, TELAM SE 2022. 
-         Essa região carece de atenção dos órgãos internacionais para fornecer apoio direto à população de interesse, ou seja, migrantes e solicitantes de asilo. Por isso, o Sustainable 
-          é uma alternativa para prestar assistência a essa população vulnerável. Um ambiente com alto índice de pobreza e o mais grave é a solidão dos migrantes sem apoio.</article>
-          <iframe width="560" height="315" src="https://www.youtube.com/embed/v0HLjtYZWPc" title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" allowfullscreen></iframe>
-      </div>
-    </section>
-
-  );
-}
-export default Estadosba;
