--- frontend/src/componentes/Footer.js
@@ -1,13 +1,0 @@
-import React from "react";
-import "./App.css";
-
-function Footer() {
-    return (
-        <div classname="text-center">
-            <p>
-                &copy; 2023 Todos os Dereitos Reservados
-            </p>
-        </div>
-    );
-}
-export default Footer;
