--- frontend/src/componentes/styles.css
@@ -1,16 +1,0 @@
-
-/*.App{
-    background-color: rgb(7, 7, 166);
-}*/
-.cars {
-  /*background-color: rgb(7, 7, 166);*/
-  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-  align-items: center;
-
-}
-Footer {
-  background-color: #ca932b;
-}
-
