--- frontend/src/componentesproducto/bellezacomponentes/roupas.css
@@ -1,20 +1,0 @@
-.produtos {
-  display: flex; /*uno al lado de otro*/
-  flex-wrap: wrap; /* os novos são empurrados para baixo*/
-  justify-content: center; /*para centralos*/
-}
-
-.produto {
-  border: 1px #404040 solid;
-  margin: 1rem; /*agregar espaço entre eles*/
-}
-
-.produto img {
-  width: 100%;
-  max-width: 280px; /*largura máxima*/
-}
-
-.produto-info {
-  text-align: left;
-  padding: 1rem;
-}
