+++ frontend/src/index.css
@@ -5,6 +5,11 @@ body {
     sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
+  /*display: flex;*/
+  /*align-items: center;*/
+  /*text-align: center;
+  /*background-color: #e0f6fc;*/
+  /* background: url("pasta/pexels-lukas-574071tres.jpg")*/
 }
 
 header {
@@ -31,17 +36,140 @@ main {
   flex: 1;
 }
 
-.btn-primary {
+.btn {
   background-color: #fdbb05;
+  color: #431a03;
+  font-weight: bold;
+}
+
+.product-name {
+  text-decoration: none;
   color: #431a03;
 }
 
 .rating span {
-  color: #fde406
+  color: #fde406;
 }
 
 .img-large {
-  max-width: 100%;
+  max-width: 55%;
 }
 
+.Empleabilidade ul li {
+  list-style: none;
+  color: white;
+}
+.Empleabilidade ul li a {
+  text-decoration: none;
+  list-style: none;
+  color: #ffda05;
+}
 
+.Empleabilidade {
+  background-color: #5d3b27;
+  color: white;
+}
+
+.produtos {
+  display: flex; /*uno al lado de otro*/
+  flex-wrap: wrap; /* os novos são empurrados para baixo*/
+  justify-content: center; /*para centralos*/
+}
+
+/*.produto {
+  border: 1px #404040 solid;
+  margin: 1rem; /*agregar espaço entre eles
+}
+
+/*.produto img {
+  width: 100%;
+  max-width: 280px; /*largura máxima
+}
+
+.produto-info {
+  text-align: left;
+  padding: 1rem;
+}*/
+
+.cars {
+  /*background-color: rgb(7, 7, 166);*/
+  display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+  align-items: center;
+}
+
+.centrasobrenos {
+  align-items: center;
+  display: flex;
+  text-align: center;
+  align-items: center;
+  justify-content: center;
+}
+
+.App-logo {
+  height: 12vmin;
+  padding: 0 1rem;
+  float: left;
+  animation: none;
+  max-width: 100%;
+  max-height: 100%;
+}
+
+.presenta_css {
+  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
+    'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
+    sans-serif;
+  margin: 0 auto;
+  width: 40vw;
+  height: 40vh;
+  padding: 30px;
+  border: 6px solid black;
+  border-radius: 4em;
+  /*background-color: #f7f7f7;*/
+  margin-top: 4em;
+}
+
+.presenta_css h3 {
+  font-size: 1.6rem;
+  padding: 0;
+  margin-bottom: 10px;
+  margin-top: 0px;
+}
+
+.presenta_css form {
+  height: 90%;
+  font-size: 1.2rem;
+  padding: 0;
+  margin: 0 auto;
+  display: flex;
+  flex-direction: column;
+  font-family: system-ui, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto,
+    Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
+  text-align: start;
+}
+
+.controls {
+  padding: 5px;
+  border-radius: 4px;
+  height: 20px;
+  border: 0.01rem solid black;
+  font-size: 1rem;
+  margin-bottom: 0.5rem;
+}
+
+.button {
+  margin-top: 1rem;
+  background-color: #ca932b;
+  color: black;
+  font-weight: bolder;
+  font-size: 1.3rem;
+  border-radius: 0.4rem;
+  border: 0.1rem solid black;
+}
+
+footer {
+  background-color: #ca932b;
+  font-weight: bold;
+  text-align: center;
+}
